requirements.txt
chess==1.10.0
Pillow==10.4.0
typer==0.12.3
rich==13.7.1
requests==2.32.3
python-dotenv==1.0.1
openai==1.51.0
matplotlib==3.9.2
